refactor: move client code and tls code to own files
